+++ src/main/java/Duke.java
@@ -1,12 +1,12 @@
-import java.util.Scanner;
-import java.util.ArrayList;
+import exception.DukeException;
+import parser.Parser;
+import storage.Storage;
+import tasks.*;
+import ui.Ui;
+
 import java.io.IOException;
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileWriter;
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
 import java.nio.file.NoSuchFileException;
 
 public class Duke {
